+++ includes/JtlConnectorAdmin.php
@@ -964,6 +964,19 @@ final class JtlConnectorAdmin
             'helpBlock' => __('Order statuses that should be imported. Default: pending, processing, on hold, completed', JTLWCC_TEXT_DOMAIN),
         ];
 
+        $paymentGateways = [];
+        if (WC()->payment_gateways() instanceof WC_Payment_Gateways) {
+            $paymentGateways = WC()->payment_gateways->payment_gateways();
+            $paymentGateways = array_combine(array_keys($paymentGateways), array_column($paymentGateways, 'title'));
+        }
+
+        $fields[] = [
+            'title' => __('Import payments with following payment types only when order is completed (usually manual payment types)', JTLWCC_TEXT_DOMAIN),
+            'type' => 'jtl_connector_multiselect',
+            'options'=> $paymentGateways,
+            'id' => Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES,
+            'value' => Config::get(Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES, Config::JTLWCC_CONFIG_DEFAULTS[Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES]),
+        ];
 
         //Add sectionend
         $fields[] = [
@@ -1807,6 +1820,9 @@ final class JtlConnectorAdmin
                 self::setupDefaultOrderStatusesToImport();
             case '1.27.1':
             case '1.28.0':
+            case '1.28.1':
+            case '1.29.0':
+                self::setupDefaultManualPaymentTypes();
             case '2.0.0':
             case '2.1.0':
             case '2.2.0':
@@ -1820,6 +1836,14 @@ final class JtlConnectorAdmin
 
         Config::updateDeveloperLoggingSettings((bool)Config::get(Config::OPTIONS_DEVELOPER_LOGGING, false));
         Config::set(Config::OPTIONS_INSTALLED_VERSION, Config::getBuildVersion());
+    }
+
+    protected static function setupDefaultManualPaymentTypes()
+    {
+        if (Config::get(Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES) === null) {
+            $paymentTypes = Config::JTLWCC_CONFIG_DEFAULTS[Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES];
+            Config::set(Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES, $paymentTypes);
+        }
     }
 
     protected static function setupDefaultOrderStatusesToImport()