+++ src/Controllers/Product/ProductAttr.php
@@ -62,7 +62,6 @@ class ProductAttr extends BaseController
         $nosearch = false;
         $fbStatusCode = false;
         $purchaseNote = false;
-        /* $fbVisibility = false;*/
         //GERMAN MARKET
         $digital = false;
         $altDeliveryNote = false;
@@ -86,8 +85,7 @@ class ProductAttr extends BaseController
                     if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_FB_FOR_WOO)) {
                         
                         if (strcmp($attrName, ProductVaSpeAttrHandler::FACEBOOK_SYNC_STATUS_ATTR) === 0) {
-                            $value = strcmp(trim($i18n->getValue()), 'true') === 0;
-                            $value = $value ? '1' : '';
+                            $value = Util::isTrue($i18n->getValue()) ? '1' : '';
                             
                             if (!add_post_meta(
                                 $productId,
@@ -107,15 +105,13 @@ class ProductAttr extends BaseController
                     }
                     if(SupportedPlugins::isActive(SupportedPlugins::PLUGIN_WOOCOMMERCE_GERMANIZED2)){
                         if($i18n->getName() === ProductVaSpeAttrHandler::GZD_IS_SERVICE) {
-                            $value = $i18n->getValue();
-                            if (in_array($value, ['yes', 'no'])) {
-                                $metaKey = '_service';
+                            $value = Util::isTrue($i18n->getValue()) ? 'yes' : 'no';
+                            $metaKey = '_service';
 
-                                if (!add_post_meta($productId, $metaKey, $value, true)) {
-                                    update_post_meta($productId, $metaKey, $value,
-                                        \get_post_meta($productId, $metaKey, true)
-                                    );
-                                }
+                            if (!add_post_meta($productId, $metaKey, $value, true)) {
+                                update_post_meta($productId, $metaKey, $value,
+                                    \get_post_meta($productId, $metaKey, true)
+                                );
                             }
                         }
                     }
@@ -155,8 +151,7 @@ class ProductAttr extends BaseController
                     if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_GERMAN_MARKET)) {
                         
                         if (strcmp($attrName, ProductVaSpeAttrHandler::GM_DIGITAL_ATTR) === 0) {
-                            $value = strcmp(trim($i18n->getValue()), 'true') === 0;
-                            $value = $value ? 'yes' : 'no';
+                            $value = Util::isTrue($i18n->getValue()) ? 'yes' : 'no';
                             $metaKey = substr($attrName, 5);
                             if (!add_post_meta(
                                 $productId,
@@ -175,8 +170,7 @@ class ProductAttr extends BaseController
                         }
                         
                         if (strcmp($attrName, ProductVaSpeAttrHandler::GM_SUPPRESS_SHIPPPING_NOTICE) === 0) {
-                            $value = strcmp(trim($i18n->getValue()), 'true') === 0;
-                            $value = $value ? 'on' : '';
+                            $value = Util::isTrue($i18n->getValue()) ? 'on' : '';
                             if ($value) {
                                 if (!add_post_meta(
                                     $productId,
@@ -235,8 +229,7 @@ class ProductAttr extends BaseController
                     }
                     
                     if (strcmp($attrName, ProductVaSpeAttrHandler::DOWNLOADABLE_ATTR) === 0) {
-                        $value = strcmp(trim($i18n->getValue()), 'true') === 0;
-                        $value = $value ? 'yes' : 'no';
+                        $value = Util::isTrue($i18n->getValue()) ? 'yes' : 'no';
                         
                         if (!add_post_meta(
                             $productId,
@@ -255,8 +248,7 @@ class ProductAttr extends BaseController
                     }
                     
                     if (strcmp($attrName, ProductVaSpeAttrHandler::PURCHASE_ONLY_ONE_ATTR) === 0) {
-                        $value = strcmp(trim($i18n->getValue()), 'true') === 0;
-                        $value = $value ? 'yes' : 'no';
+                        $value = Util::isTrue($i18n->getValue()) ? 'yes' : 'no';
                         
                         if (!add_post_meta(
                             $productId,
@@ -275,8 +267,7 @@ class ProductAttr extends BaseController
                     }
                     
                     if (strcmp($attrName, ProductVaSpeAttrHandler::VIRTUAL_ATTR) === 0) {
-                        $value = strcmp(trim($i18n->getValue()), 'true') === 0;
-                        $value = $value ? 'yes' : 'no';
+                        $value = Util::isTrue($i18n->getValue()) ? 'yes' : 'no';
                         
                         if (!add_post_meta(
                             $productId,
@@ -296,7 +287,7 @@ class ProductAttr extends BaseController
                     }
                     
                     if ($attrName === ProductVaSpeAttrHandler::PAYABLE_ATTR || $attrName === 'payable') {
-                        if (strcmp(trim($i18n->getValue()), 'false') === 0) {
+                        if (Util::isTrue($i18n->getValue()) === false) {
                             \wp_update_post([
                                 'ID'          => $productId,
                                 'post_status' => 'private',
@@ -304,16 +295,16 @@ class ProductAttr extends BaseController
                             $payable = true;
                         }
                     }
-
+                    
                     if ($attrName === ProductVaSpeAttrHandler::NOSEARCH_ATTR || $attrName === 'nosearch') {
-                        if (strcmp(trim($i18n->getValue()), 'true') === 0) {
+                        if (Util::isTrue($i18n->getValue())) {
                             \update_post_meta(
                                 $productId,
                                 '_visibility',
                                 'catalog',
                                 \get_post_meta($productId, '_visibility', true)
                             );
-
+                            
                             /*
                             "   exclude-from-catalog"
                             "   exclude-from-search"
@@ -413,9 +404,9 @@ class ProductAttr extends BaseController
                 );
             }
         }
-
+        
         if (!$nosearch) {
-
+            
             if (!\add_post_meta(
                 $productId,
                 '_visibility',
@@ -431,7 +422,7 @@ class ProductAttr extends BaseController
             }
             \wp_remove_object_terms($productId, ['exclude-from-search'], 'product_visibility');
         }
-
+        
         if (!$purchaseNote) {
             if (!\add_post_meta(
                 $productId,