+++ src/Utilities/Config.php
@@ -26,6 +26,7 @@ class Config
         OPTIONS_VARIATION_NAME_FORMAT = 'jtlconnector_variation_name_format',
         OPTIONS_DEFAULT_CUSTOMER_GROUP = 'jtlconnector_default_customer_group',
         OPTIONS_DEFAULT_ORDER_STATUSES_TO_IMPORT = 'jtlconnector_default_order_statuses_to_import',
+        OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES = 'jtlconnector_default_manual_payment_types',
         OPTIONS_USE_GTIN_FOR_EAN = 'jtlconnector_use_gtin_for_ean',
         OPTIONS_ALLOW_HTML_IN_PRODUCT_ATTRIBUTES = 'jtlconnector_allow_html_in_product_attributes',
         OPTIONS_COMPLETED_ORDERS = 'jtlconnector_completed_orders',
@@ -47,6 +48,7 @@ class Config
         Config::OPTIONS_ALLOW_HTML_IN_PRODUCT_ATTRIBUTES => false,
         Config::OPTIONS_DEFAULT_CUSTOMER_GROUP => 'customer',
         Config::OPTIONS_DEFAULT_ORDER_STATUSES_TO_IMPORT => ['wc-pending', 'wc-processing', 'wc-on-hold'],
+        Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES => ['cod', 'german_market_purchase_on_account', 'german_market_sepa_direct_debit', 'cheque', 'bacs', 'cash_on_delivery', 'invoice'],
         //PAGE
         Config::OPTIONS_USE_DELIVERYTIME_CALC => true,
         Config::OPTIONS_DISABLED_ZERO_DELIVERY_TIME => true,
@@ -73,6 +75,7 @@ class Config
         Config::OPTIONS_ALLOW_HTML_IN_PRODUCT_ATTRIBUTES => 'bool',
         Config::OPTIONS_DEFAULT_CUSTOMER_GROUP => 'string',
         Config::OPTIONS_DEFAULT_ORDER_STATUSES_TO_IMPORT => 'array',
+        Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES => 'array',
         //PAGE
         Config::OPTIONS_USE_DELIVERYTIME_CALC => 'bool',
         Config::OPTIONS_DISABLED_ZERO_DELIVERY_TIME => 'bool',