+++ src/Utilities/Util.php
@@ -494,6 +494,16 @@ final class Util extends WordpressUtils
     }
 
     /**
+     * @return array
+     */
+    public static function getManualPaymentTypes(): array
+    {
+        $defaultManualPayments = Config::JTLWCC_CONFIG_DEFAULTS[Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES];
+        return Config::get(Config::OPTIONS_DEFAULT_MANUAL_PAYMENT_TYPES, $defaultManualPayments);
+    }
+
+
+    /**
      * @param string $stateName
      * @return bool
      */
@@ -617,5 +627,14 @@ final class Util extends WordpressUtils
     public static function getStates()
     {
         return WC()->countries->get_states();
+    }
+
+    /**
+     * @param string $value
+     * @return bool
+     */
+    public static function isTrue(string $value): bool
+    {
+        return !in_array(strtolower(trim($value)), ['no', '0', 'false', ''], true);
     }
 }